Implement third-party addon support core
This commit implements the necessary core features to allow us
to add third-party addon support later.

It adds a callback-system where our third party addon code
can register callbacks that will run on the creation of the
main-spellbuttons, the rank-spellbuttons and on the main
window's show- and hide-events. This will allow us to deeply
integrate any third party addons that we want to support,
while still keeping that support-code separate from the
actual TinyBook code, thanks to the clean callback system.